0.60

+++ scripts/cli_unfreezing.sh
@@ -18,10 +18,14 @@
 # https://github.com/swtor00/swtor-addon-to-tails       #
 #########################################################
 
+if [ -f ~/Persistent/swtorcfg/freezed.cgf ] ; then
+   rm -rf /live/persistence/TailsData_unlocked/dotfiles/.config > /dev/null 2>&1
+   rm -rf /live/persistence/TailsData_unlocked/dotfiles/Desktop > /dev/null 2>&1
+   rm -rf /live/persistence/TailsData_unlocked/dotfiles/Pictures > /dev/null 2>&1
 
-rm -rf /live/persistence/TailsData_unlocked/dotfiles/.config > /dev/null 2>&1
-rm -rf /live/persistence/TailsData_unlocked/dotfiles/Desktop > /dev/null 2>&1
-rm -rf /live/persistence/TailsData_unlocked/dotfiles/Pictures > /dev/null 2>&1
+   rm ~/Persistent/swtorcfg/freezed.cgf > /dev/null 2>&1
+else
+   echo "unfreezing is not possible. This system is not freezed"
+fi
 
-rm ~/Persistent/swtorcfg/freezed.cgf > /dev/null 2>&1
